requirements.txt
requests
click
ffmpeg-python
srt
